+++ apps/backoffice-v2/src/lib/blocks/components/PDFViewerCell/interfaces.ts
@@ -1,0 +1,10 @@
+export interface IPDFViewerProps {
+  width: string;
+  height: string;
+}
+
+export type TPDFViewerCell = {
+  type: 'pdfViewer';
+  value: string;
+  props: IPDFViewerProps;
+};