+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.1-11210907c.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.0-11210907c.0
+  - @ballerine/workflow-core@0.6.1-11210907c.0
+
 ## 0.6.0
 
 ### Minor Changes